requirements.txt
aiofiles==22.1.0
bleach==5.0.1
blinker==1.5
caldav==0.10.0
certifi==2022.9.24
chardet==5.0.0
charset-normalizer==2.1.1
click==8.1.3
DateTimeRange==1.2.0
h11==0.14.0
h2==4.1.0
hpack==4.0.0
hypercorn==0.14.3
hyperframe==6.0.1
icalendar==5.0.1
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
lxml==4.9.1
MarkupSafe==2.1.1
mbstrdecoder==1.1.1
packaging==21.3
priority==2.0.0
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==0.21.0
pytz==2022.5
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
quart==0.18.3
requests==2.28.1
six==1.16.0
toml==0.10.2
typepy==1.3.0
tzdata==2022.6
tzlocal==4.2
urllib3==1.26.12
uuid==1.30
vobject==0.9.6.1
webencodings==0.5.1
Werkzeug==2.2.2
wsproto==1.2.0
